requirements.txt
streamlit
streamlit-webrtc
pandas
pillow
opencv-python-headless
exifread
numpy

